Added axios requests/request generator
